requirement.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
networkx==3.1
nltk==3.8.1
scipy==1.10.1
joblib==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.0
plotly==5.15.0